Add style to third element on top of existing ones
The third element already has a class value using which we have applied
styles to odd numbered elements. Now we want to apply additional styles
specific to only the third element on top of the currently applied odd
numbered element's style. For this requirement we can add a new id
unique to this third element on the page and then applying styles to it
or we can add a new class to the third element by separating them with
the whitespace and then applying style specific to this class value.
Also since this third element already has an existing class value added
to it we can select this element using the chaining selector.