Add style to proportioned and avatar class image
To apply style to specific element which has a class which is shared
with other elements we use the chaining selector to add those specific
styles. So here to add the specific style to image element which has
class value containing classes "avatar" and "proportioned" we use the
chaining selector.