Add styles for header UI using margin auto in flex
Added the styles for creating the UI for the header using the flexbox
and for spreading out the items on both used the trick of margin auto
value instead of the justify content property. For justify-content we
would have to create addition containers around item groups which needs
to be pushed on either sides.