Add style for creating the holy grail page layout
Added styles for creating a page layout consisting the header, content
and footer. In the content section we have a non-shrinkable sidebar on
the left and on the right we have the main content which is the cards
which wraps onto the new line when there is no space available
horizontally.